Few fixes with Tetra quests (need more info)
I did as much as I could. In advancements there are still only 6 tiers of Hammer progression, and its the same as in quest. I moved those "artifacts" to other section however, and added note that forge hammer was just deleted, not "not yet implemented"
Sorry @mickelus, but I'm not able to do anything more atm ^^"